Coding Days: grid generators
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2736 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/generation/PrewaveletGridGenerator.hpp
@@ -38,8 +38,13 @@ private:
 public:
 	/**
 	 * Constructor
+	 * An adaptive grid with prewavelet ansatz functions requires for operations
+	 * using the up-down algorithm shadow points. These shadow points a needed just
+	 * for data transport, thus they do not have an influence on the final function.
+	 * Please refer to sg::pde::UpDownOneOpDimWithShadow for more information.
 	 *
 	 * @param storage the grid's storage object
+	 * @param shadowstorage shadow points (see detailed description)
 	 */
 	PrewaveletGridGenerator(GridStorage* storage, GridStorage* shadowstorage);
 
@@ -48,8 +53,8 @@ public:
 	 */
 	virtual ~PrewaveletGridGenerator();
 
-	virtual void regular(size_t level);
-	virtual void full(size_t level);
+	virtual void regular(int level);
+	virtual void full(int level);
 	virtual void refine(RefinementFunctor* func);
 	virtual size_t getNumberOfRefinablePoints();
 
@@ -57,8 +62,8 @@ public:
 	virtual void coarsenNFirstOnly(CoarseningFunctor* func, DataVector* alpha, size_t numFirstOnly);
 	virtual size_t getNumberOfRemovablePoints();
 
-	virtual void refineMaxLevel(RefinementFunctor* func, unsigned int maxLevel);
-	virtual size_t getNumberOfRefinablePointsToMaxLevel(unsigned int maxLevel);
+	virtual void refineMaxLevel(RefinementFunctor* func, int maxLevel);
+	virtual size_t getNumberOfRefinablePointsToMaxLevel(int maxLevel);
 
 };
