+++ src/sgpp/base/grid/generation/StretchedTrapezoidBoundaryGridGenerator.hpp
@@ -35,8 +35,8 @@ public:
 	 */
 	virtual ~StretchedTrapezoidBoundaryGridGenerator();
 
-	virtual void regular(size_t level);
-        virtual void full(size_t level);
+	virtual void regular(int level);
+        virtual void full(int level);
 	virtual void refine(RefinementFunctor* func);
 	virtual size_t getNumberOfRefinablePoints();
 
@@ -44,8 +44,8 @@ public:
 	virtual void coarsenNFirstOnly(CoarseningFunctor* func, DataVector* alpha, size_t numFirstOnly);
 	virtual size_t getNumberOfRemovablePoints();
 
-	virtual void refineMaxLevel(RefinementFunctor* func, unsigned int maxLevel);
-	virtual size_t getNumberOfRefinablePointsToMaxLevel(unsigned int maxLevel);
+	virtual void refineMaxLevel(RefinementFunctor* func, int maxLevel);
+	virtual size_t getNumberOfRefinablePointsToMaxLevel(int maxLevel);
 
 protected:
 	/// Pointer to the grid's storage object
